- added last resort built-in mediatype for capture filter if card does not returns any (i.e Osprey cards) - miscelaneous clean ups

+++ modules/access/dshow/filter.h
@@ -209,7 +209,7 @@ class CaptureEnumMediaTypes : public IEnumMediaTypes
     access_t * p_input;
     CapturePin     *p_pin;
 
-    int i_position;
+    size_t i_position;
     long i_ref;
 
 public:
